WIP remove optional height indexing interface
