+++ chains/linearizable_vm.go
@@ -17,9 +17,8 @@ import (
 )
 
 var (
-	_ vertex.LinearizableVM      = (*initializeOnLinearizeVM)(nil)
-	_ block.ChainVM              = (*linearizeOnInitializeVM)(nil)
-	_ block.HeightIndexedChainVM = (*linearizeOnInitializeVM)(nil)
+	_ vertex.LinearizableVM = (*initializeOnLinearizeVM)(nil)
+	_ block.ChainVM         = (*linearizeOnInitializeVM)(nil)
 )
 
 // initializeOnLinearizeVM transforms the consensus engine's call to Linearize
@@ -63,15 +62,12 @@ func (vm *initializeOnLinearizeVM) Linearize(ctx context.Context, stopVertexID i
 // channel to the VM that is being linearized.
 type linearizeOnInitializeVM struct {
 	vertex.LinearizableVMWithEngine
-	hVM          block.HeightIndexedChainVM
 	stopVertexID ids.ID
 }
 
 func NewLinearizeOnInitializeVM(vm vertex.LinearizableVMWithEngine) *linearizeOnInitializeVM {
-	hVM, _ := vm.(block.HeightIndexedChainVM)
 	return &linearizeOnInitializeVM{
 		LinearizableVMWithEngine: vm,
-		hVM:                      hVM,
 	}
 }
 
@@ -87,20 +83,4 @@ func (vm *linearizeOnInitializeVM) Initialize(
 	_ common.AppSender,
 ) error {
 	return vm.Linearize(ctx, vm.stopVertexID, toEngine)
-}
-
-func (vm *linearizeOnInitializeVM) VerifyHeightIndex(ctx context.Context) error {
-	if vm.hVM == nil {
-		return block.ErrHeightIndexedVMNotImplemented
-	}
-
-	return vm.hVM.VerifyHeightIndex(ctx)
-}
-
-func (vm *linearizeOnInitializeVM) GetBlockIDAtHeight(ctx context.Context, height uint64) (ids.ID, error) {
-	if vm.hVM == nil {
-		return ids.Empty, block.ErrHeightIndexedVMNotImplemented
-	}
-
-	return vm.hVM.GetBlockIDAtHeight(ctx, height)
 }
