+++ snow/engine/snowman/block/vm.go
@@ -5,10 +5,16 @@ package block
 
 import (
 	"context"
+	"errors"
 
 	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/avalanchego/snow/consensus/snowman"
 	"github.com/ava-labs/avalanchego/snow/engine/common"
+)
+
+var (
+	ErrHeightIndexedVMNotImplemented = errors.New("vm does not implement HeightIndexedChainVM interface")
+	ErrIndexIncomplete               = errors.New("query failed because height index is incomplete")
 )
 
 // ChainVM defines the required functionality of a Snowman VM.
@@ -48,6 +54,22 @@ type ChainVM interface {
 	// a definitionally accepted block, the Genesis block, that will be
 	// returned.
 	LastAccepted(context.Context) (ids.ID, error)
+
+	// VerifyHeightIndex should return:
+	// - nil if the height index is available.
+	// - ErrIndexIncomplete if the height index is not currently available.
+	// - Any other non-standard error that may have occurred when verifying the
+	//   index.
+	VerifyHeightIndex(context.Context) error
+
+	// GetBlockIDAtHeight returns:
+	// - The ID of the block that was accepted with [height].
+	// - database.ErrNotFound if the [height] index is unknown.
+	//
+	// Note: A returned value of [database.ErrNotFound] typically means that the
+	//       underlying VM was state synced and does not have access to the
+	//       blockID at [height].
+	GetBlockIDAtHeight(ctx context.Context, height uint64) (ids.ID, error)
 }
 
 // Getter defines the functionality for fetching a block by its ID.
