--- vms/tracedvm/height_indexed_vm.go
@@ -1,39 +1,0 @@
-// Copyright (C) 2019-2023, Ava Labs, Inc. All rights reserved.
-// See the file LICENSE for licensing terms.
-
-package tracedvm
-
-import (
-	"context"
-
-	"go.opentelemetry.io/otel/attribute"
-
-	oteltrace "go.opentelemetry.io/otel/trace"
-
-	"github.com/ava-labs/avalanchego/ids"
-	"github.com/ava-labs/avalanchego/snow/engine/snowman/block"
-)
-
-func (vm *blockVM) VerifyHeightIndex(ctx context.Context) error {
-	if vm.hVM == nil {
-		return block.ErrHeightIndexedVMNotImplemented
-	}
-
-	ctx, span := vm.tracer.Start(ctx, vm.verifyHeightIndexTag)
-	defer span.End()
-
-	return vm.hVM.VerifyHeightIndex(ctx)
-}
-
-func (vm *blockVM) GetBlockIDAtHeight(ctx context.Context, height uint64) (ids.ID, error) {
-	if vm.hVM == nil {
-		return ids.Empty, block.ErrHeightIndexedVMNotImplemented
-	}
-
-	ctx, span := vm.tracer.Start(ctx, vm.getBlockIDAtHeightTag, oteltrace.WithAttributes(
-		attribute.Int64("height", int64(height)),
-	))
-	defer span.End()
-
-	return vm.hVM.GetBlockIDAtHeight(ctx, height)
-}
